move accessors to access package
